% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{explode_graph}
\alias{explode_graph}
\title{Explode discourse graph (make statement nodes qualified)}
\usage{
explode_graph(disc_g, stance_subset = SUPPORTED_STANCE_CLASSES)
}
\arguments{
\item{disc_g}{A discourse graph object}

\item{stance_subset}{Character vector of stance classes. Restrict explosion to a specific set of stance classes only? Defaults to all supported.}
}
\description{
This function "explodes" a discourse graph, making all statement nodes into qualified nodes.
}
