% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{get_igraph}
\alias{get_igraph}
\title{Extract igraph object from discourse graph}
\arguments{
\item{g}{The discourse graph to extract igraph object from}
}
\value{
An object of class igraph
}
\description{
Extract igraph object from discourse graph
}
