% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{get_incmat}
\alias{get_incmat}
\title{Get incidence matrix of discourse graph}
\usage{
get_incmat(disc_g, make_binary = FALSE)
}
\arguments{
\item{disc_g}{A discourse graph object}

\item{make_binary}{Logical. Dichotomize/ make binary? Defaults to FALSE, which returns a valued matrix with the number of stances in entries.}
}
\description{
Get incidence matrix of discourse graph
}
