% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{get_tbl_graph}
\alias{get_tbl_graph}
\title{Extract tidygraph object from discourse graph}
\arguments{
\item{g}{The discourse graph to extract tidygraph object from}
}
\value{
An object of class tidygraph
}
\description{
Extract tidygraph object from discourse graph
}
