% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{is.discourse_graph}
\alias{is.discourse_graph}
\title{Check if object is of type discourse_graph}
\usage{
is.discourse_graph(object)
}
\arguments{
\item{disc_g}{}
}
\value{
boolean
}
\description{
Check if object is of type discourse_graph
}
\examples{
is.discourse_graph(diskurs::discourse_graph_example)
}
