% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_discourse_graph}
\alias{load_discourse_graph}
\title{Load discourse graph from nodelist of actors/ statements and edgelist of expressed stances}
\usage{
load_discourse_graph(nodelist, edgelist)
}
\arguments{
\item{nodelist}{A data frame containing the following columns:
nodeid: integer vector of unique node identifiers
name: a unique name assigned to a node
label: the label of the node
mode: one of 'actor' or 'statement'}

\item{edgelist}{An edgelist specifying qualified relations between actors and statements.
Takes the form of a data frame containing the following columns:
from: actor node ids expressing stances
to: statement node ids toward which statements are expressed
stance: One of 'support', 'opposition' or 'irrelevant', capturing the qualifier for the stance expression
timestamp: A timestamp given the time the stance was expressed}
}
\value{
A discourse graph object
}
\description{
Load discourse graph from nodelist of actors/ statements and edgelist of expressed stances
}
\examples{
diskurs::load_discourse_graph(diskurs::nodelist_example,diskurs::edgelist_example)
}
