% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot}
\alias{plot}
\title{Plot discourse graph}
\arguments{
\item{disc_g}{A discourse graph object}

\item{layout_choice}{Layout choice to pass to ggraph. Defaults to "kk"}

\item{label_nodes}{Logical: should nodes be labelled? Defaults to TRUE}

\item{edge_alpha}{Transparency of edges. Defaults to 1}

\item{arrow_start_cap}{Space from arrow start cap. Defaults to 6.}

\item{arrow_end_cap}{Space from end of arrow. Defaults to 10.}
}
\description{
Plot discourse graph
}
