% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{remove_isolates}
\alias{remove_isolates}
\title{Remove isolates from discourse graph object}
\usage{
remove_isolates(disc_g)
}
\arguments{
\item{disc_g}{A discourse graph object}
}
\value{
A discourse graph object with isolated nodes (actors with no expressed stances or statements with no refering stances) removed
}
\description{
Remove isolates from discourse graph object
}
