% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{time_slice_graph}
\alias{time_slice_graph}
\title{Time-slice graph}
\usage{
time_slice_graph(disc_g, start_date, end_date)
}
\arguments{
\item{disc_g}{Discourse graph object}

\item{start_date}{The beginning of the time slice in a format convertible by lubridate::as_date}

\item{end_date}{The end of the time slice in a format convertible by lubridate::as_date}
}
\value{
A discourse graph object subset to include only edges within the time slice given by start_date and end_date
}
\description{
Time-slice graph
}
