% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{time_sliced_graph_list}
\alias{time_sliced_graph_list}
\title{Create a list of time-sliced graphs}
\usage{
time_sliced_graph_list(
  disc_g,
  time_window,
  date_range,
  step_interval = "month",
  aggregate = FALSE,
  remove_isolates = FALSE,
  keep_only_highest = FALSE
)
}
\arguments{
\item{disc_g}{A discourse graph object}

\item{time_window}{The slice window as a time period object, eg. created with months(48)}

\item{date_range}{Character vector of length 2 with a start and end date coercible to date by lubridate::as_date()}

\item{step_interval}{The step interval for time slices. Defaults to "month".}

\item{aggregate}{Logical. Aggregate the time sliced graphs? Defaults to FALSE.}

\item{remove_isolates}{Logical. Remove isolates per time slices? Defaults to FALSE.}

\item{keep_only_highest}{Logical. If aggregate is TRUE, keep only most prevalent stance category per time slice? Defaults to FALSE.}
}
\description{
Create a list of time-sliced graphs
}
